Report WebRTC Git subtree revision for Chromium builders.
The reason r_webrtc_git is not used is that in a Chromium checkout, the
src/third_party/webrtc dir is still synced from a Git mirror, so using
the https://chromium.googlesource.com/external/webrtc/ URL for blame lists
will render invalid links.
It's probably possible to implement parsing of the Cr-Mirrored-Commit Git
footer but since we'll try to get rid of the subtree mirror I prefer not to:
the important thing here is to give a working blame list link to the perf
sheriffs.

BUG=491520,482330,382370

Change-Id: Ia4d5e5a5635a3a77fb98ec4552601b85ba5f6fc5
TBR=kjellander@chromium.org
Reviewed-on: https://chromium-review.googlesource.com/415791
Commit-Queue: Henrik Kjellander <kjellander@chromium.org>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>