added a support for "make status" for run_slave.
In Linux, /etc/init.d/buildbot status returns an error, and, therefore,
Service['buildbot'] in puppet cannot detect the running status of buildbot
service. This CL is to control buildbot service in puppet with Service {}.

Tested on a windows bot with test.pp to verify the feature.

R=dsansome
TBR=friedman
BUG=601499

Change-Id: I4ed84ebef0eb70f314772c588980283e5626b978
Reviewed-on: https://chromium-review.googlesource.com/416508
Reviewed-by: Scott Lee <ddoman@google.com>
Reviewed-by: Dave Sansome <dsansome@chromium.org>
Commit-Queue: Scott Lee <ddoman@google.com>