Fixing layout of abort reason, removing line breaks from exception.
The details of why a bisect was inconclusive (such as a range of
revisions failing to build) are both the abort reason to be specified
in the bisect report, and the details of the bisect failure. However,
for failing a bisect, we need a string with no newlines as a message
for the StepFailure exception, while for readability we need them in
the bisect report. This change removes the newlines from the message
before raising the failure.

BUG=665439
TBR=simonhatch@chromium.org

Change-Id: I1c49040675d3cb7346392590e4ff1278404e06fa
Reviewed-on: https://chromium-review.googlesource.com/417127
Reviewed-by: Roberto Carrillo <robertocn@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>