Use the Chromium-side script to run the PGO benchmarks (reland).
This uses this script: https://cs.chromium.org/chromium/src/build/win/run_pgo_profiling_benchmarks.py
instead of the one that was in the resources/ directory (the scripts are
the same). The recipes are not pinned on the official builders, so it is
currently not possible to update the benchmark set without affecting all
the milestones that we still ship, and having the benchmark set living
on the Chromium side will make it easier to measure the impact of a change
to this set (by using a perf trybot).

This is a reland of https://codereview.chromium.org/2421363002/ which had
to be reverted because it wasn't the appropriate values for the
'--target-cpu' flag (should be x86 or x64)

BUG=309849

Change-Id: I24574d01a309b0d4ca31ec321b1f2978faffa7d9
Reviewed-on: https://chromium-review.googlesource.com/419255
Commit-Queue: Sébastien Marchand <sebmarchand@chromium.org>
Reviewed-by: Sébastien Marchand <sebmarchand@chromium.org>