use goma module in findit recipes
This is retry of following CL
https://chromium-review.googlesource.com/c/412504/

findit recipes assume that if there is a failure in goma, compile.py stop and raise exception.
But in default config, compile.py disables goma and continues build
when goma fails (even if it becomes too slow).

If there is a goma related failure, it is better to detect goma unexpected behavior as soon as possible.
So I enable goma_failfast in this config.

BUG=656540

Change-Id: I3c94f73efdba044f90804f757785b8d34a0463c1
Reviewed-on: https://chromium-review.googlesource.com/420183
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Takuto Ikuta <tikuta@google.com>