Adds specific compile targets to the ChromiumOS codesearch builders.
We currently use the 'all' target which is causing compile problems for
code search that go unnoticed on the ChromiumOS waterfall. This solution
lists the targets currently used by 'Linux ChromiumOS Full' builder. This
connects the code search builder to the waterfal for the moment, but the
target lists can easily become out of sync.

The target used determines the coverage of cross-references in code search.
A failing build gives us nothing, and no config gives us all, so this is
best effort.

BUG=674850

Change-Id: I4374bcf805bc6ea597824596d0f33e08390156dd
Reviewed-on: https://chromium-review.googlesource.com/421506
Commit-Queue: Emma Söderberg <emso@chromium.org>
Reviewed-by: Paweł Hajdan Jr. <phajdan.jr@chromium.org>