Re-add support for additional_compile_targets to iOS recipes.
Support for specifying 'additional_compile_targets' in an iOS recipe
was removed in https://chromium-review.googlesource.com/c/416320/ .
I think this was accidental; I noticed it in my review but stupidly
forgot to ask about it. However, the recipes should generally
support this because it is functionality that is needed from time to
time when you want to be sure a recipe builds something that isn't a
test.

This is needed at the moment as part of the iOS upstreaming work, so
I'm adding it back in. I will double-check w/ smut@ when they're
available in case they wanted to approach this some other way and will
change it if so.

R=sdefresne@chromium.org
TBR=smut@chromium.org
BUG=598405, 653086

Change-Id: I77366c92109a719f9978cba702053952b193b740
Reviewed-on: https://chromium-review.googlesource.com/421182
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@chromium.org>