Modify the isolate recipe to support exparchive isolates.
We're testing a new method for archiving very large isolates, and
we need to be able to isolate some targets using the `exparchive`
command instead of the `batcharchive` command. This will hopefully
be a temporary thing until the functionality is merged into the
main `archive` and `batcharchive` commands, and so for simplicity
the recipe will just decide that any target ending in '_exparchive'
should use the experimental mode.

R=maruel@chromium.org, djd@chromium.org, mcgreevy@chromium.org
BUG=524758

Change-Id: Id04f3e0daa9543c8c40e67f60530a2b1a145abe1
Reviewed-on: https://chromium-review.googlesource.com/422460
Commit-Queue: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Michael McGreevy <mcgreevy@chromium.org>