Add support for flakiness testing GTests.
Add a new recipe that runs a test a number of times (100) by default for a single revision
and reports the number of times the test passes/fails.

This change supports GTests both local and swarmed and reuses as much code from test.py as
possible.

R=stgao@chromium.org, phajdan.jr@chromium.org, iannucci@chromium.org
BUG=672721

Change-Id: I646ec35657b841fdc8ae4270aac6a19ca4b2d84d
Reviewed-on: https://chromium-review.googlesource.com/418875
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Paweł Hajdan Jr. <phajdan.jr@chromium.org>