WebRTC: Fix gtest-parallel argument passing and enable parallelism for modules_tests in client.webrtc.fyi
If -- is inserted after gtest-parallel and the test binary, the flags after that 
token will be passed to each parallel execution of the binary, which is not what
we want for our current buildbot setup. Possibly in the future, if tests require
non-gtest flags to configure their runs, this might be necessary.

BUG=webrtc:6946
TBR=ehmaldonado@chromium.org

Change-Id: I646f590001ac9d5eb9bc117af9896773c9db0505
Reviewed-on: https://chromium-review.googlesource.com/424394
Commit-Queue: Henrik Kjellander <kjellander@chromium.org>
Reviewed-by: Henrik Kjellander <kjellander@chromium.org>