Create a new logs column in test results table.
Previously tombstones can be viewed if clicking on 'CRASH' in the status
column and there is a separate column for logcat.

In this cl, we remove the tombstones from the status column and remove
the logcat column. Instead we create a new 'logs' column, which has both
tombstones and logcat as url links.

BUG=605572

Change-Id: I1148eb7f637f9a105955aea07333947a940c2c5f
Reviewed-on: https://chromium-review.googlesource.com/424144
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Zhiling Huang <hzl@google.com>