gae_sdk updater no longer uses VERSION file.
The GS bucket used to have a VERSION file that contains, among other
things, the latest version of the AppEngine SDK. That file still exists,
but now contains "0.0.0" as the current version.

Rather than rely on that file, instead list the contents of the bucket
for each platform and identify the latest version by parsing the
structured filename.

BUG=None
TEST=expectations

Change-Id: I6a25e313ae1e109cb8ecfcc5202dcabc2ad501ae
Reviewed-on: https://chromium-review.googlesource.com/426407
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Daniel Jacques <dnj@chromium.org>