Migrate chromium.gpu.fyi waterfall to builders.pyl.
The port was done by going through slaves.cfg and creating new entries
one-by-one.

The only significant difference between this and the previous configuration
is that the VM testers (which run their jobs on Swarming) are pooled per
platform. This reduced the number of slave pools that had to be defined.

BUG=646162

Change-Id: I68a9451f956e12734dba857ebbcc9ba9046cc282
Reviewed-on: https://chromium-review.googlesource.com/426960
Commit-Queue: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>