TBR=dsansome@chromium.org
Revert "Adds Gold upload support to the pdfium recipe. "

This reverts commit 61e29b63bd53227c31062f672e8d766ddbf6dcc6.

Error in Commit 61e29b6 - need to revert the last 3 changes.

The issue is that "calendar" module isn't whitelisted, as shown in the CQ Error message.

"calendar" module will need to be whitelisted in "infra/recipes-py/recipe_engine/lint_test.py"

List of changes reverted ...

 1 a506fb2 (most recent)
 2 c6ff1f3
 3 61e29b6 (contains error)


This is revert 3 of 3

Change-Id: I0d5b98f6407f5551f38c99d782617094b23b6139
Reviewed-on: https://chromium-review.googlesource.com/428653
Reviewed-by: Phil Wright <philwright@chromium.org>