Fix bug in chartjson upload code
It incorrectly attempts to upload chartjson, even if there is no data,
and the bot isn't correctly configured to upload data, it will try to
upload something. This doesn't cause any errors currently, because of a
try catch Exception around the whole upload block, but it is a bug.

This CL bails out of uploading if it detects that the bot isn't
correctly configured to upload chart json.

BUG=680840

Change-Id: Ia68fbb47f3db4aaa480f2d812bd0de57f9fb25ff
Reviewed-on: https://chromium-review.googlesource.com/428180
Reviewed-by: Mike Stipicevic <stip@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>