Buildbot PubSub: Keep track of all pending builds in the pubsub modules
This uses a variety of subscription services within to keep track of pending builds.

Buildbot backs pending build reqeusts by an SQL database for the case of multiple
masters trying to claim the same request.  But since we don't use multiple
masters with a single database, this is not a concern for us and it's okay to
maintain all pending builds in memory.

BUG=673202

Change-Id: Iff288ed9dc39317f4c0e6f1ab61429228e1b91b6
Reviewed-on: https://chromium-review.googlesource.com/427228
Reviewed-by: Daniel Jacques <dnj@chromium.org>
Commit-Queue: Ryan Tseng <hinoka@chromium.org>