Switch master.chromium.gpu to builders.pyl.
This mainly copied over the already-working bots from
master.chromium.gpu.fyi, adjusting the machines used.

This loses the gatekeeper categories for the GPU builders, but this
doesn't seem like a big problem at this point. The tryservers are
already preventing breakage of the GPU targets.

This is the last reference to recipe_master_helper. Delete it.

BUG=646162

Change-Id: I9ecb8142612a20ced9dc2c5bb3c36f0b9b174c67
Reviewed-on: https://chromium-review.googlesource.com/430253
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Kenneth Russell <kbr@chromium.org>