Create link in test results presentation page to report bugs/issues.
Test results presentation is a good way to see why a test fails.
Unfortunately it didn't provide a way to quickly create bugs for
developers to triage.

In this cl, a report issues/bugs link is created at the bottom which
will redirect to the bug creation website with several fields populated
already.

BUG=605572

Change-Id: Ib497673b98564db74b6e164c637abe32bf3b1526
Reviewed-on: https://chromium-review.googlesource.com/430215
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Zhiling Huang <hzl@google.com>