Tee stdout for perf bisect tests (non-staging bots).
Buildbot is timing out since we intercept stdout and buildbot
does not see any output for more than 60 minutes. Tee'ing output
so we can intercept it and it is streamed to buildbot.

Landed this CL for staging_* bisect bots. Would like to now
land this change for the rest of the bots.

Change-Id: I892616e8f2b1688afdc4e1fdca62b7f8f8bc2411

BUG=664765

Change-Id: I892616e8f2b1688afdc4e1fdca62b7f8f8bc2411
Reviewed-on: https://chromium-review.googlesource.com/430239
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Michael Case <mikecase@chromium.org>