chromium_tests: raise an error on inconsistent swarming config
More specifically, fail builders which disable swarming on builder level,
but have tests in testing/buildbot that seem to enable swarming (but it's
still effectively disabled on builder level).

The goal of this CL is to catch all of current occurrences of this confusion
(which leads to build failures when swarming gets unexpectedly enabled
on these builders) and prevent regressions.

BUG=684527, 684067

Change-Id: Ibdcbaa1b8e0611c8e197110a75e953570b6b9dfd
Reviewed-on: https://chromium-review.googlesource.com/430682
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Paweł Hajdan Jr. <phajdan.jr@chromium.org>