Add max failures for SwarmingGTestTest
Add a property allowing specification of a maximum number of test
failures. Below this number, the recipe will retry the step without the
patch. Above this number, the build will fail.

We're adding this because if a test suite has an extremely large number
of failures, there is likely an issue, and we don't want to waste time
retrying the tests. Additionally, this should take care of the "extra
long gtest filter string" errors we've seen occasionally.


BUG=667004

Change-Id: I3e16201e86ef97790f873059c9d1227371254929
Reviewed-on: https://chromium-review.googlesource.com/430155
Reviewed-by: Dirk Pranke <dpranke@chromium.org>