Remove LogDog-less retries.
LogDog bootstrapping has traditionally responded to bootstrap failures
(either due to configuration or due to infrastructure-attributed
failures in the LogDog or Annotee binaries) by re-executing the build
sans the LogDog bootstrap.

This CL makes it so that if a LogDog bootstrap is configured, it MUST
execute, and any failure will result in a build failure.

This is important since the fallback of sending logs to BuildBot is no
longer acceptable for BuildBot masters that rely on "logdog_only" mode
for security.

BUG=chromium:685843
TEST=unit

Change-Id: Id9c1eb524a799dcee029e01e003e49eb96ad0841
Reviewed-on: https://chromium-review.googlesource.com/433368
Commit-Queue: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>