Rename incorrectly-named Mac GPU bots.
Many of these bots included a now-incorrect OS version in their names.
Remove it and clarify some of the other bots' names.

While this renaming will lose build history, these bots have been
green for a while, so now is as good a time as any to rename them.

This depends on the Chromium-side patch
https://codereview.chromium.org/2679543002 which provides buildbot
JSON files for the new machine names.

BUG=688887
TBR=dpranke@chromium.org,phajdan.jr@chromium.org
Change-Id: Icdb831c17750cdd5b937190e1e283e82f0be3065

Change-Id: Icdb831c17750cdd5b937190e1e283e82f0be3065
Reviewed-on: https://chromium-review.googlesource.com/437616
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Kenneth Russell <kbr@chromium.org>