Add a custom postprocess/merge hook and wrapper for isolated script tests.
This adds a way to customize how the output from swarming tasks
is merged. It also allows you to run client-side work after the
tasks are completed and the results are collected, and to optionally
customize the links on the collect step in the buildbot page.

These things will be used by the layout tests to merge the
results JSON and archive the results html dirs.

This CL was migrated from dpranke's original at
https://chromium-review.googlesource.com/c/423114/

R=maruel@chromium.org, nednguyen@chromium.org, dpranke@chromium.org
BUG=676210

Change-Id: Ie502920549b434782238b6f7cc9e006bb3cab66b
Reviewed-on: https://chromium-review.googlesource.com/433397
Commit-Queue: John Budorick <jbudorick@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Tim 'mithro' Ansell <tansell@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>