Add a custom postprocess/merge hook and wrapper for isolated script tests. (RELAND)
This adds a way to customize how the output from swarming tasks
is merged. It also allows you to run client-side work after the
tasks are completed and the results are collected, and to optionally
customize the links on the collect step in the buildbot page.

These things will be used by the layout tests to merge the
results JSON and archive the results html dirs.

This CL is a reland of https://chromium-review.googlesource.com/c/433397/

BUG=676210

Change-Id: I6fbff9955d5c269f039d026d1451aa9447f79f69
Reviewed-on: https://chromium-review.googlesource.com/441391
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: John Budorick <jbudorick@chromium.org>