Reactor logs: Use __call__ instead of args[0]
This is the third try of logging something useful about the reactor state.
First try: f is a callable, so lets print the name.  They mostly turned out to
           be named "callback", so we thought they were deferred.
Second try: If f is always a deferred, then arg[0] must be the deferred object.
            This turned out not to be the case.

Now print the __call__ tag, which looks something like this:
<code object b at 0x7f4eb7e60bb0, file "a.py", line 4>

This should be more useful.

BUG=

Change-Id: Id47de3d3bbfa5dcf60b467425b9285799cc77b55
Reviewed-on: https://chromium-review.googlesource.com/443764
Reviewed-by: Daniel Jacques <dnj@chromium.org>
Commit-Queue: Ryan Tseng <hinoka@chromium.org>