Add the component mapping updater bot to gatekeeper.
The bot periodically runs a script that traverses src/
and examines OWNERS files and extracts the `# TEAM:` 
and `# COMPONENT:` tags in those files into a mapping
that is then uploaded (if modified) to:
https://storage.googleapis.com/chromium-owners/component_map.json.

We are doing this on a tryserver because that's where we had the
capacity. After some time, we expect to move it to a more
approriate location.

BUG=688579
R=​martiniss

Change-Id: Ifd51c008fabb1f94646894c942e5e5a04b95b8cd
Reviewed-on: https://chromium-review.googlesource.com/437614
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>