LogDog-only mode for tryserver.chromium.linux
As a diagnostic measure, reduce (~eliminate) the logging data that the
tryserver has to handle. The hope is that this will make it more
responsive and solve the immediate problem in the bug.

This mode has been running on Infra and Perf waterfalls for over a month
without issue, so this should be a safe change.

BUG=chromium:692708
TEST=None

Change-Id: I80bd9e25f08544439446aebe73286b810f015513
Reviewed-on: https://chromium-review.googlesource.com/445917
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Sergey Berezin <sergeyberezin@chromium.org>
Commit-Queue: Daniel Jacques <dnj@chromium.org>