Add remote_run test, canarying, pin recipes
Add canarying to "remote_run". Pin the "recipes-py" package
version so that arbitrary updates don't lose the basic ability to launch
recipes.

Additionally, add a much-needed suite of "remote_run" tests that,
similar to "annotated_run", assert the correctness of basic execution
and expected command-line flags.

Some refactoring was needed to accommodate the unit test approach and
canarying. Some additional refactoring is probably present due to the
fact that this was backported from its successor patch, which introduces
Kitchen. The expected output behavior of this patch is that "remote_run"
behavior remains unchanged, except for the pins.

BUG=chromium:694911
TEST=local

Change-Id: I7197aa4cbe0b3578933b6f53371e9d85535fabe7
Reviewed-on: https://chromium-review.googlesource.com/446904
Commit-Queue: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Dave Sansome <dsansome@chromium.org>