Emit a "log_location" protobuf when logdog-only.
Builds that are in logdog-only mode will now emit a "log_location"
property specifying the LogDog annotation URL. This corresponds with a
similar mechanism in SwarmBucket.

The short-term goal is to enable BuildBot templates to identify
logdog-only mode by looking for this property and checking the prefix
against "logdog://".

BUG=chromium:698375
TEST=unit

Change-Id: I248e12ef849b3cb6aaa48fd08eca596a18a9bf4c
Reviewed-on: https://chromium-review.googlesource.com/448965
Reviewed-by: Ryan Tseng <hinoka@chromium.org>
Commit-Queue: Daniel Jacques <dnj@chromium.org>