Explicitly pass gsutil.py path to runtest.
"runtest.py" expects to use depot_tools. In hermetic recipe
enviornments, the path to depot_tools is not inferrable from the
path of the recipe's working directory like it is in traditional
BuildBot environments. Fix this by explicitly passing the path of
"gsutil.py" to "runtest.py".

BUG=chromium:698984
TEST=expectations

Change-Id: I0b917a6e08dfbc3e303f15711ccfc15e0d8ff332
Reviewed-on: https://chromium-review.googlesource.com/450955
Commit-Queue: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>