[test_results] remove __init__.py from resources
This is inadvertantly causing the recipe engine to load all of the
resources python scripts into the recipe engine when this module is
imported. Normally this isn't an issue, but because scripts in this
directory do sys.path manipulation, it means that any recipes that
import this module (directly or indirectly) will end up with a goofy
PYTHONPATH that leaks down to all steps in the recipe.

Removing this means that e.g.
  `import RECIPE_MODULES.test_results.resources.test_results_uploader`

Won't work anymore. However that's insane and nothing does that :)

BUG=

Change-Id: I0d02bb1862f092cb65da1b277d146df784f3ba11
Reviewed-on: https://chromium-review.googlesource.com/451882
Reviewed-by: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Paweł Hajdan Jr. <phajdan.jr@chromium.org>
Reviewed-by: Sergiy Byelozyorov <sergiyb@chromium.org>
Commit-Queue: Paweł Hajdan Jr. <phajdan.jr@chromium.org>