PubSub: Stop the buildbot master if we can't init a client
As we move towards using the pubsub stream for all data, we need to make sure
the startup sequence works, and doesn't just fail silently.  This treats
the failure to obtain a PubSub client as a critical failure in the master
(if the client is configured to run in master_site_config).

BUG=699625

Change-Id: Iaf3e7c13d66ccad74c710fd605b40f9dcb89857f
Reviewed-on: https://chromium-review.googlesource.com/451797
Commit-Queue: Ryan Tseng <hinoka@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Dave Sansome <dsansome@chromium.org>