[Cronet] Fix sizes.py to ignore version number in libcronet.so
libcronet.so is now named to include the version, for example
libcronet.58.0.3028.0.so.  This change makes the sizes.py script
find the newly named binary but strip the version from the name
when printing results so we comparison scripts can continue to
simply compare libcronet.so and historical results are included.
This gets sizes.py working again on the Cronet builders after
crrev.com/455971

R=xunjieli

Change-Id: I1c4a513aa2239c914a32e919920651293fe23ba3
Reviewed-on: https://chromium-review.googlesource.com/452323
Reviewed-by: John Budorick <jbudorick@chromium.org>
Reviewed-by: Xunjie Li <xunjieli@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Paul Jensen <pauljensen@chromium.org>