Export master and slave parameters from run_slave
Currently, "chromium_utils" offers the ability to probe the active
master and slave name. This relies on legacy directory layouts, and
offers the opportunity for these values to be out of sync if the
directory used in the probe falls behind HEAD (which is what is running
on the actual bot).

Have "run_slave.py", the actual BuildBot agent, export this information.
It is in a position of authority to do so. Modify "chromium_utils" to
look for these exports and, if available, prefer them to its own probing.

Review needs to be CRITICAL, since a bug here will result in
irrecoverably offline bots.

BUG=chromium:700823
TEST=local
  - Ran `pylint`, passes.

Change-Id: Ie21a5554ee4c5e88210136caf7bb3172c18b4c72
Reviewed-on: https://chromium-review.googlesource.com/453919
Commit-Queue: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>