Simplify layout_test_wrapper.py (take 2).
This is a second attempt of simplifying
layout_test_wrapper.py in order to prepare to remove it.

Previous CL was:
https://chromium-review.googlesource.com/c/452548/

In this version, all command-line arguments are forwarded on
to run-layout-tests, but some of the extra things that the
wrapper currently does are left alone in this version,
including cleaning up results and temporary directories,
and removing padding from and checking validity of
the failing_results.json file.

BUG=695700

Change-Id: Ie8290e8202a0f4c93f518e3ef40f5ba06e59ac90
Reviewed-on: https://chromium-review.googlesource.com/456402
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Tim 'mithro' Ansell <tansell@chromium.org>