Make AOT simarm64 bot use blobs instead of dylibs.
We've been using the host (x64) toolchain to generate a dylib with arm64 code it in, and we've been getting away with it because because all the instructions were pre-encoded. But now we're trying to include architecture specific unwind information, so this trick won't work anymore.

Switch the AOT linux bots to use dylibs to maintain coverage.

Change-Id: I36be75df6520506d6ff6ff83de032e0bd43285ef
Reviewed-on: https://chromium-review.googlesource.com/456589
Reviewed-by: Vyacheslav Egorov <vegorov@chromium.org>
Commit-Queue: Vyacheslav Egorov <vegorov@chromium.org>