WebRTC: Use test name instead of command for perf_dashboard_id fallback.
The low_bandwidth_audio_test is the first perf test that is using a Python
script to execute. Previously we've only had simple test executables, which
have always been valid Perf IDs. A full path to a script is not, which is
why I'm changing the fallback when perf_dashboard_id isn't specified.

BUG=webrtc:7229
TBR=ehmaldonado@chromium.org

Change-Id: I5d44745826d0d7f6e863d9aa41cdf2446fc0fe41
Reviewed-on: https://chromium-review.googlesource.com/458405
Reviewed-by: Henrik Kjellander <kjellander@chromium.org>
Commit-Queue: Henrik Kjellander <kjellander@chromium.org>