Apply LogDog-only rendering to all views (#3).
This is a reland of: https://chromium-review.googlesource.com/c/458986/

Currently, LogDog-only rendering is implemented only for the "build"
view. This makes clicks on the waterfall and console dysfunctional,
since they will link to the useless LogDog-only notice messages instead
of the real logs.

Rather than painfully updating every site and template, this introduces
a generic projection function that transforms a BuildStep's log and URL
entries into LogDog-only compatible entries when appropriate. All
applicable build status sites have been updated to use the
transformation funciton.

This is not expected to introduce any rendering performance degradation,
since all of the functions that it uses are already called as part of
rendering.

This supercedes the functionality introduced in:
https://chromium-review.googlesource.com/450019

That CL can be safely reverted after this one lands.

TBR=hinoka@chromium.org
BUG=chromium:701218
TEST=local
  - Visited all views w/ a local BuildBot instance, verified that
    everything seems to be in order.

Change-Id: I24b57feb4ac32dc04c9a95be6bdc96d515ebba56
Reviewed-on: https://chromium-review.googlesource.com/459445
Reviewed-by: Daniel Jacques <dnj@chromium.org>
Commit-Queue: Daniel Jacques <dnj@chromium.org>