Make sure that step presentation is not set to FAILURE if exit_code is zero
In crbug.com/680138, we introduce 'ignore_swarming_task_failure' which allows
swarming isolated script tests to express that it's ok for it to fail.

This is a follow up after https://chromium-review.googlesource.com/458261 to
ensure that if the test's exit_code is zero, step presentation will not be set
to FAILURE even if there are unexpected failures.


BUG=680138

Change-Id: Id44491241918be6bd0c8acd85e9f9df995a96614
Reviewed-on: https://chromium-review.googlesource.com/462407
Reviewed-by: Tim 'mithro' Ansell <tansell@chromium.org>
Commit-Queue: Ned Nguyen <nednguyen@google.com>