Fix the step link of result details.
Previouly the link of result details is not showing up in the waterfall.

The reason of the failure is that the link we provide contains two
lines, because 'print' in python will automatically add new line char at
the very end.

In this cl, I am stripping the new line char at the very end.

Bug:

Change-Id: I0fdf3cf0209dd176b07fae2a84fbbe32990d1a10
Reviewed-on: https://chromium-review.googlesource.com/466607
Reviewed-by: Michael Case <mikecase@chromium.org>
Commit-Queue: Zhiling Huang <hzl@google.com>