Fix recipe invocation of mb analyze on mac to check FORCE_MAC_TOOLCHAIN.
The filter recipe_module code that runs the "analyze" logic calls out
to mb.py directly to call analyze, rather than call into the chromium
module's run_mb() method. As a result, there is code duplication, and
there was a line in the chromium method that wasn't in the filter
method, and we were incorrectly not exporting the FORCE_MAC_TOOLCHAIN
variable.

R=erikchen@chromium.org
TBR=phajdan.jr@chromium.org
NOTREECHECKS=true
BUG=708657

Change-Id: Id842967d32c935aa7750a9209e7f672e93d4bb19
Reviewed-on: https://chromium-review.googlesource.com/469076
Commit-Queue: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>