Run "cleanup_temp" during build setup.
Currently, the "cleanup_temp" script is run by recipes explicitly as a
way of purging previous state from the system. This sort of thing is not
something that individual recipes should concern themselves with;
rather, the system launcher should ensure a sufficiently clean
environment as a prerequisite to running the recipe in the first place.

Consequently, we will move the "cleanup_temp" invocation into the recipe
launcher areas, "annotated_run.py" and "remote_run.py". This will ensure
that recipes are executed in a pristine environment.

This will be followed with CLs to remove "cleanup_temp" invocations from
all recipes, under the assumption that the cleanup is now done for them.

BUG=chromium:708616
TEST=None

Change-Id: I66db240bb98ccd38f6a8bbfb300759f5bb3375c0
Reviewed-on: https://chromium-review.googlesource.com/475064
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Daniel Jacques <dnj@chromium.org>