Add option to look up the triggered tests from another master.
This allows the tryserver.chromium.perf compilers to build and isolate the same targets as their corresponding chromium.perf ones.

We only have an option to override the master, not the builder name, because it would be counterintuitive to simulate a builder with a different name, and Pinpoint requires they have the same name anyway.

Bug: catapult:#3394
Change-Id: Ib842d9510576dc4c7d6c88d081ba420aea8bff29
Reviewed-on: https://chromium-review.googlesource.com/475109
Commit-Queue: David Tu <dtu@chromium.org>
Reviewed-by: Annie Sullivan <sullivan@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>