Seal off "annotated/remote_run" unit tests.
Currently, the "annotated_run" and "remote_run" unit tests
AnnotatedRunTest and RemoteRunTest respectively run out of the active build
directory. This can be problematic, since it operates on that and parent
directories with the assumption that they are actually live builder
directories, including the creation and management of "/b/..."
directories.

The tests determine "/b" by walking up from their current working
directory. We can, therefore, use a fake build directory by creating a
temporary "<TEMP>/b/build/slave/buildername/build" directory and setting it
as the current working directory when running the subordinate scripts.

This should affect the problem in the linked bug.

We also update "update_scripts" (RISKY!) to allow the directory that it
syncs to be specified via environment variable. We add a
FakeBuildRootTestCase unit test which creates a fake test root and
exports this environment variable, compelling "update_scripts" to update
the test directory instead of the actual checkout. This should maintain
the standard of asserting the correctness of "update_scripts" while
containing it.

BUG=chromium:713506
TEST=local

Change-Id: If1d0090fdb9877f426f9ce34163d4accd0ee5fa8
Reviewed-on: https://chromium-review.googlesource.com/483383
Commit-Queue: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>