Disable kitchen on Site Isolation Linux for now.
kitchen sets TEMPDIR and TMPDIR in order to attempt to make sure
that its subprocesses are only writing to directories it can manage.
However, if you run Xvfb, it appears to want to create something
that exists in a hard-coded /tmp/foo location, and so it seems to
try and create a hard link from /tmp/foo to TEMPDIR/foo. Which
doesn't work when TEMPDIR is on a different filesystem.

This CL disables kitchen for now, until we can come up with a fix.

TBR=dnj@chromium.org, tansell@chromium.org
BUG=708681, 715848

Change-Id: I54700551de824c8c6fc8ec3d5e4d74ee24c260eb
Reviewed-on: https://chromium-review.googlesource.com/488133
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@chromium.org>