Don't force PGO builds to use VS 2015
Currently PGO builds are hard-coded to use VS 2015 - they ignore the
default compiler as set in build\vs_toolchain.py, which leads to
confusion. For instance, the 60.0.3086.0 canary used VS 2015 and this
was initially assumed to be because it started building after the VS
2017 default-change was reverted, but that is not actually the case.
VS 2017 became the default on #468239 and stopped being the default
on #468298, but the canary build was built from #468266.

Because the default compiler is currently VS 2015 this should not make
any difference until the default in build\vs_toolchain.py is changed.

The CL that changed the default compiler from VS 2013 to VS 2015
(crrev.com/1607063004) was more complex because it included changes to
paths, whereas switching between VS 2015 and VS 2017 should be a one
character change.

Bug: 683729
Change-Id: Idbde58fc9649f5021d60c594a5a71df940a73c34
Reviewed-on: https://chromium-review.googlesource.com/492108
Reviewed-by: Sebastien Marchand <sebmarchand@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>