Fetch Telemetry binary dependencies on main Linux and Mac builders.
This has the effect of downloading these during the runhooks step on
the Linux and Mac builders on the chromium.linux, chromium.mac and
chromium.gpu waterfalls, as well as on the linux_chromium_rel_ng and
mac_chromium_rel_ng tryservers.

These dependencies are already being downloaded and included in the
isolates on the chromium.gpu.fyi waterfall, and the mechanism is
working well and has delivered tremendous speedups on some of the
tests.

Also enabled this on the ASAN and TSAN builders on chromium.gpu.fyi.

BUG=711307

Change-Id: Ib931572d874c24c2c91f719bbd76ec98748c4c40
Reviewed-on: https://chromium-review.googlesource.com/495635
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Kenneth Russell <kbr@chromium.org>