Merge build requests on testers on the chromium.chromiumos waterfall.
The testers were configured to take every triggered build from
the builders and run the tests; this means that if the tests take
longer than the builders, we would become farther and farther behind,
which is what was happening.

This CL changes things to coalesce (merge) the triggered requests
so that we can keep up.

R=agable@chromium.org
BUG=718287

Change-Id: I7409a4f07f04846fb1931b80840e647540d4d7af
Reviewed-on: https://chromium-review.googlesource.com/496726
Reviewed-by: Aaron Gable <agable@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@chromium.org>