Revert Kitchen bump.
The Kitchen bump (correctly) made it so that parent process PYTHONPATH
was propagated into the child process. Unfortunately, this broke Glyco
tests, since that test relies on "setuptools" and the propagated
PYTHONPATH includes a broken version of "setuptools". Revert the roll
until a more permanent fix can be made.

Realistically this should have been a blocking problem with respect to
turning Kitchen on for this waterfall in the first place. The Kitchen
bug where it stomped PYTHNONPATH masked the issue, and fixing the bug
revealed it.

BUG=None
TEST=None
R=bpastene@chromium.org, nodir@chromium.org

Change-Id: Ia15e69d98ac2aed2c7b3981e2d3f0a61597e53cf
Reviewed-on: https://chromium-review.googlesource.com/498832
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Daniel Jacques <dnj@chromium.org>