remote_run: Decouple canary from Kitchen.
Decouple canary configuration from Kitchen configuration. This allows
canarying Recipes/Kitchen packages independently from the decision about
whether or not to actually use Kitchen. Masters that are considered
canary masters are tracked in a separate list.

Allow Kitchen configuration to be expressed through blacklists in
addition to whitelists. The goal of the Kitchen experiment is to migrate
builders over to Kitchen and, if successful, not have them migrate back.
On some large masters, this can be better expressed through a blacklist
of builders that don't work (to burn down) as opposed to a whitelist of
builders that do work.

BUG=chromium:694911
TEST=unit

Change-Id: I4d93aea7ab482cd27760f884bb2175b14ec52c67
Reviewed-on: https://chromium-review.googlesource.com/498709
Commit-Queue: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>