Fix bug in recipe result handling logic
It turns out that remote_run and annotated_run had a bug in how they
handled the engine flags. This CL fixes it, and adds tests.

The code in annotated_run isn't using this logic at the moment, so no
change is expected there. remote_run behavior will potentially change,
in the case of infrastructure failures. This will correctly set the exit
code of remote_run. Right now, it exits with code 1, when it really
should be exiting with code 255. An example build is here:
https://uberchromegw.corp.google.com/i/chromium.perf/builders/Android%20Nexus5%20Perf%20%281%29/builds/5582/steps/steps/logs/stdio
The build still looks like a infrastructure failure, so looks like this
bug had minimal impact.

Change-Id: I6dfaae98b457890c2c202264ba7c45cdebb03304
Reviewed-on: https://chromium-review.googlesource.com/489426
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Daniel Jacques <dnj@chromium.org>